[Front] Unified BuildLogs and DockerFiles stuff, with small css fixes
