--- front/src/components/BuildLogs/index.js
@@ -1,1 +1,0 @@
-export { default } from './BuildLogs';
