--- front/src/components/DockerFiles/index.js
@@ -1,1 +1,0 @@
-export { default } from './DockerFiles';
