+++ front/src/components/LinesBlocks/LinesBlocks.css
@@ -7,18 +7,22 @@ h3 {
   margin-bottom: 0.5em;
 }
 
+.block {
+  --atom-one-dark-background: 40, 44, 51;
+  background-color: rgb(var(--atom-one-dark-background));
+  padding: 1.2em 0;
+}
+
 .lines {
   --atom-one-dark-background: 40, 44, 51;
-  max-height: 15em;
+  max-height: 25em;
   white-space: pre-wrap;
   font-family: monospace;
   overflow-y: scroll;
   background-color: rgb(var(--atom-one-dark-background));
-  padding: 0 1.2em !important;
-  color: white;
-  border: 1.2em solid rgb(var(--atom-one-dark-background));
-  border-left: none;
-  border-right: none;
+  margin: 0;
+  padding: 0 1.2em;
+  color: #eeeeee;
   font-size: 0.8em;
 }
 
@@ -26,9 +30,6 @@ h3 {
   display: block;
 
   & + & {
-    margin-top: 0.5em;
-  }
-  &::before {
-    content: "$ ";
+    margin-top: 0.2em;
   }
 }
