+++ front/src/components/LinesBlocks/index.js
@@ -1,0 +1,1 @@
+export { BuildLogs, DockerFiles } from './LinesBlocks';
